-first come first serve & Round robin working in graph form => komt overeen met powerpoint resulaten , graph begint gwn niet bij null maar hij zei dat dat geen kwaad kan gwn punten verbinden is goed.
-logartmischa as toegevoegd voor TAT
-Result Processor rekent nu met doubles ipv integers (accurater)

+++ src/practicum1/DataProcessing/Containers/GraphData.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 
 //List of 100 values for generating an XY Graph
 
-public class GraphData extends ArrayList<Integer> {
+public class GraphData extends ArrayList<Double> {
 
 
     private String title;
